Updates Node.js version and manages dependencies
Updates the default Node.js version to v20.9.0 for general use
and installs Node.js v17.9.1 for compatibility with the server-api.

This ensures the 'lib' directory is built with the latest Node.js,
while maintaining compatibility for the 'server-api' directory,
which depends on an older Node.js version.

Also updates npm to v10.0.0 globally.