Tests updates to work with callout refresh (#441)
* Tests updates to work with callout refresh

* Address CR feedback

* Update

* Check the error

* fix node version

* Fixed and enabled storage tests

* Updates Node.js version and manages dependencies

Updates the default Node.js version to v20.9.0 for general use
and installs Node.js v17.9.1 for compatibility with the server-api.

This ensures the 'lib' directory is built with the latest Node.js,
while maintaining compatibility for the 'server-api' directory,
which depends on an older Node.js version.

Also updates npm to v10.0.0 globally.

* Updates Travis CI configuration

Uses Ubuntu 20.04 and Node.js 20.11.0 for better compatibility.

This change ensures the Travis CI environment is compatible and
uses a more recent Node.js version.

* Removes npm update from Travis CI config

Removes the global npm update from the Travis CI configuration, as it is no longer necessary. The required npm version is now included with the newer Node version.

* Removes sourceMap option from tsconfig

Removes the `sourceMap` option from the tsconfig.json file and relies on `inlineSourceMap` and `inlineSources` instead.

This simplifies the configuration and reduces the number of files generated during the build process.

* Replaces inline source map generation

Updates the tsconfig to generate separate source map files instead of inlining them.
This change improves debugging workflow by providing better source code mapping in development environments.

---------

Co-authored-by: Svetoslav <svetoslav@alkem.io>
Co-authored-by: Valentin Yanakiev <valentin.yanakiev@gmail.com>